+++ code/com.zengqiangfang/src/main/java/com/zengqiangfang/service/CalculatorForPpiService.java
@@ -12,7 +12,7 @@ import org.springframework.stereotype.Service;
  */
 @Service
 public class CalculatorForPpiService {
-    public long calculate(int width, int height, int size){
+    public long calculate(int width, int height, double size){
         if(width > 0 && height > 0 && size > 0){
             return Math.round(Math.pow((Math.pow(width, 2) + Math.pow(height, 2))/ size, 0.5));
         }
